float

+++ training/training_loop.py
@@ -183,8 +183,8 @@ def training_loop(
 
             def __next__(self):
                 img, labels = next(self.iterator)
-                img = img.type(torch.HalfTensor).to(device)
-                # img = img.type(torch.FloatTensor).to(device)
+                # img = img.type(torch.HalfTensor).to(device)
+                img = img.type(torch.FloatTensor).to(device)
                 encoded = autoencoder.encode(img)
                 del img
                 return encoded, labels
